TODO: Rewrite some services, Location problem

+++ backend/Unicorn.Core/Interfaces/IAccountService.cs
@@ -7,6 +7,6 @@ namespace Unicorn.Core.Interfaces
     public interface IAccountService
     {
         Task<IEnumerable<AccountDTO>> GetAllAsync();
-        Task<AccountDTO> GetById(int id);
+        Task<AccountDTO> GetById(long id);
     }
 }
