+++ backend/Unicorn.Core/Interfaces/IPersonService.cs
@@ -7,6 +7,6 @@ namespace Unicorn.Core.Interfaces
     public interface IPersonService
     {
         Task<IEnumerable<PersonDTO>> GetAllAsync();
-        Task<PersonDTO> GetById(int id);
+        Task<PersonDTO> GetById(long id);
     }
 }
